gpu-direct: Fix build of test/HD/KHI due to cyclic includes + fix missing #if WITH_MPI when MPI is disabled
