+++ src/utils/column.hpp
@@ -59,7 +59,7 @@ class Column {
   IdefixArray3D<real> Area;
   IdefixArray3D<real> Volume;
 
-    idefix::IdefixCommArray2D<real> localSum;
+  idefix::IdefixCommArray2D<real> localSum;
   #ifdef WITH_MPI
   Mpi mpi;  // Mpi object when WITH_MPI is set
   MPI_Comm ColumnComm;
